pyproject.toml
[tool.poetry]
name = "django-movie-recommendation"
version = "0.1.0"
description = ""
authors = ["Diego Margoni <diegof.margoni@@gmail.com>"]
readme = "README.md"
packages = [{include = "django_movie_recommendation"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
django-redis = "^5.2.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.2.0"
black = "^23.3.0"
django-silk = "^5.0.3"
ipdb = "^0.13.13"
factory-boy = "^3.2.1"
pre-commit = "^3.2.2"
pre-commit-hooks = "^4.4.0"
pyupgrade = "^3.3.2"
pylint = "^2.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
